OrientationModule added & camera improvments
Camera movement is a bit buggy with the addition of the z direction

Planets orbits seem to be a bit slower than expected. Moon is behind earth for the supposed eclipse date and the earths orbit is 2 months longer than expected. Will need to check previous versions to see if this was always the case or not. Could be orbital decay as a result of RK4